cleaned up and added some TODOs

+++ a1/common/src/commonMain/kotlin/at/ac/tuwien/foop/common/models/dtos/socket/PositionDto.kt
@@ -9,7 +9,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class PositionDto(
     @SerialName("x")
-    var x: Long,
+    var x: Int,
     @SerialName("y")
-    var y: Long,
+    var y: Int,
 )
